Massive re-arrange for Maven support. DO NOT MERGE OR YOU WILL BREAK YOUR BUILD

+++ org.openntf.domino.xsp/src/org/openntf/domino/xots/annotations/Schedule.java
@@ -15,7 +15,7 @@ public @interface Schedule {
 
 	TimeUnit timeunit();
 
-	int frequency();
+	long frequency();
 
 	int[] weekdays() default { ALL };
